pyproject.toml
[tool.poetry]
name = "langchain-rag-tutorial-intermediary"
version = "0.1.0"
description = ""
authors = ["gmcosta <geu_costa@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.76"
pytest = "^8.3.4"
chromadb = "^0.5.23"
langchain = "^0.3.10"
pypdf = "^5.1.0"
langchain-openai = "^0.2.11"
langchain-community = "^0.3.10"
python-dotenv = "^1.0.1"
langchain-chroma = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
